requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-admin-autocomplete-filter==0.7.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
